Remove ArrayDefinitionBuilder
